Added Convention rights

+++ src/main/java/com/eval/demo/controller/ConventionController.java
@@ -67,9 +67,9 @@ public class ConventionController {
 
     // DELETE
     @JsonView(ConventionView.class)
-    @IsAdmin
+    @IsEntreprise
     @DeleteMapping("/convention/{id}")
-    public ResponseEntity<Convention> deleteConvention(@PathVariable Integer id) {
+    public ResponseEntity<Convention> deleteConvention(@PathVariable Integer id, @AuthenticationPrincipal AppUserDetails appUserDetails) {
 
         //On vérifie que l'Convention existe bien dans la base de donnée
         Optional<Convention> optionalConvention = conventionDao.findById(id);
@@ -77,6 +77,11 @@ public class ConventionController {
         //si l'convention n'existe pas
         if(optionalConvention.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        if(!appUserDetails.getUtilisateur().getDroit().equals("ADMINISTRATEUR") &&
+                !optionalConvention.get().getEntreprise().getUtilisateur().getId().equals(appUserDetails.getUtilisateur().getId())) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
 
         conventionDao.deleteById(id);
